% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_total_binomial}
\alias{beta_total_binomial}
\title{beta_total_binomial}
\format{
A list containing:
\itemize{
\item \code{beta_target}: numeric vector of true coefficients.
}
}
\source{
Simulation
}
\description{
Simulated coefficient vector for the binomial case.
}
