% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_total_linear}
\alias{beta_total_linear}
\title{beta_total_linear}
\format{
A list containing:
\itemize{
  \item \code{beta_target}: numeric vector of true coefficients.
}
}
\source{
Simulation
}
\usage{
beta_total_linear
}
\description{
Simulated coefficient vector for the linear case.
}
\keyword{datasets}
