\name{CatlGLM-package}
\alias{CatlGLM-package}
\alias{CatlGLM}
\title{CatlGLM: compositional adaptive transfer learning}
\description{
Package containing tools for adaptive transfer learning with compositional
covariates under zero-sum constraints for linear and logistic regression,
plus de-biased confidence intervals.
}
